Revamped kernel: made it target arch independant.
The goal was to make the kernel free of any target architecture.
The one-architecture-only code has been moved in the arch folder, and
is not compiled if the architecture doesn't match.

Moved x86-only code to arch/x86.
Moved linker script to arch/$(ARCH)
Moved include/asm.h to include/arch/$(ARCH)/
Moved dev/ to arch/$(ARCH)/devices/
Deleted C code about the gdt, moved that directly in the boot code in assembly.
Edited stuff here and there in the makefile and scripts to support multiple
target architectures.